arreglos

+++ funciones.h
@@ -1,19 +1,19 @@
 /**------------------------------------------------ESTRUCTURAS--------------------------------------------------------**/
 struct Arbol{
-	float valor;
+	double valor;
 	struct Arbol *izq;
 	struct Arbol *der; 
 };
 /**----------------------------------------------FIN ESTRUCTURAS-------------------------------------------------------**/
 
-long abrirArchivoArreglo(float **arreglo,FILE *archivo);
+long abrirArchivoArreglo(double **arreglo,FILE *archivo);
 void abrirArchivoArbol(struct Arbol **arbol);
 
-void merge(float *arreglo,long inicio,long mitad,long final);
-void mergeSort(float *arreglo,long inicio,long final);
+void merge(double *arreglo,long inicio,long mitad,long final);
+void mergeSort(double *arreglo,long inicio,long final);
 
-int particion(float *arreglo,int menor,int mayor);
-void quicksort(float *arreglo,int menor,int mayor);
+int particion(double *arreglo,int menor,int mayor);
+void quicksort(double *arreglo,int menor,int mayor);
 
-void imprimirArreglo(float *arreglo,long largo);
-int estaOrdenado(float *arreglo,long largo);
+void imprimirArreglo(double *arreglo,long largo);
+int estaOrdenado(double *arreglo,long largo);
